Struct fireplace::Config

pub struct Config {
    pub outputs: HashMap<String, OutputConfig>,
    pub logging: Logging,
    pub keys: HashMap<String, KeyPattern>,
    pub view: View,
    pub exec: Exec,
    pub workspace: WorkspacesConfig,
    pub focus: FocusConfig,
    pub screenshot: ScreenshotConfig,
}

Main configuration struct

Collects all configuration structs from the various handlers. Can be deserialized by serde.

Fields
outputs: HashMap<String, OutputConfig>

A HashMap of an OutputConfig for each name of an Output. The name default or generic might be used for a general configuration.

logging: Logging

Logging configuration

keys: HashMap<String, KeyPattern>

A HashMap of global actions that may be invoked through keys.

terminate => End the compositor

view: View

Configuration of keys related to Views

exec: Exec

Configuration of program execution by keys.

workspace: WorkspacesConfig

Configuration for Workspaces

focus: FocusConfig

Configuration for View focusing

screenshot: ScreenshotConfig

Configuration for Screenshots